backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pdfplumber>=0.11.0
spacy>=3.7.0
nltk>=3.8.0
sentence-transformers>=2.7.0
scikit-learn>=1.4.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
ollama>=0.2.0
numpy>=1.26.0

# ── Auth dependencies (Phase 2) ──────────────────────────────────────────────
pyotp>=2.9.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
aiosmtplib>=3.0.0
redis>=5.0.0
qrcode[pil]>=7.4.0
pydantic[email]>=2.7.0
httpx>=0.27.0
